either expect a reason or just print an overview

+++ orstatus-stats.sh
@@ -2,13 +2,23 @@
 # set -x
 
 
-# create from input files a histogram like:
-#     0       417
-#     1       1065
-#     2       355
+# work on output of orstatus.py, eg.:
 #
-# read it as:
-#   417 nodes had 0 <reason>, 1065 had 1 <reason>, 355 had 2 <reason>, ...
+# $> orstatus.py --ctrlport 29051 >> /tmp/orstatus.29051 &
+#
+# wait a while
+#
+# $> orstatus-stats.sh /tmp/orstatus.29051
+#     146 CONNECTRESET
+#     337 DONE
+#     539 IOERROR
+#       1 NOROUTE
+#      36 TIMEOUT
+#     137 TLS_ERROR
+#
+# plot a histogram:
+#
+# $>  orstatus-stats.sh /tmp/orstatus.29051 TIMEOUT
 
 
 #######################################################################
@@ -16,28 +26,40 @@ set -eu
 export LANG=C.utf8
 export PATH="/usr/sbin:/usr/bin:/sbin:/bin:/opt/tb/bin"
 
-reason=${1?reason needed}
-shift
-[[ $# -gt 0 ]]
+files=""
+reason=""
+for opt in $*
+do
+  if [[ -e $opt ]]; then
+    files+=" $opt"
+  else
+    reason=$opt
+  fi
+done
 
-tmpfile=$(mktemp /tmp/$(basename $0)_XXXXXX.tmp)
-grep -h "^$reason " $* |\
-awk '{ print $2 }' | sort     | uniq -c |\
-awk '{ print $1 }' | sort -bn | uniq -c |\
-awk '{ print $2, $1 }' | tee $tmpfile
+if [[ -z $reason ]]; then
+  # just count reasons
+  awk '{ print $1 }' $files | sort | uniq -c
+else
+  tmpfile=$(mktemp /tmp/$(basename $0)_XXXXXX.tmp)
+  grep -h "^$reason " $files |\
+  awk '{ print $2 }' | sort     | uniq -c |\
+  awk '{ print $1 }' | sort -bn | uniq -c |\
+  awk '{ print $2, $1 }' | tee $tmpfile
 
-xmax=$(tail -n 1 $tmpfile | awk '{ print $1 }')
-(( xmax = xmax + 2))
+  xmax=$(tail -n 1 $tmpfile | awk '{ print $1 }')
+  (( xmax = xmax + 2))
 
-gnuplot -e '
-  set terminal dumb;
-  set title "fingerprints";
-  set key noautotitle;
-  set xlabel "ioerrors";
-  set xrange [-0.5:'"$xmax"'];
-  set yrange [0.5:*];
-  set logscale y 10;
-  plot "'$tmpfile'" pt "o";
-  '
+  gnuplot -e '
+    set terminal dumb;
+    set title "fingerprints";
+    set key noautotitle;
+    set xlabel "ioerrors";
+    set xrange [-0.5:'"$xmax"'];
+    set yrange [0.5:*];
+    set logscale y 10;
+    plot "'$tmpfile'" pt "o";
+    '
 
-rm $tmpfile
+  rm $tmpfile
+fi
